moved propertychanged base
