+++ Sublist.Common/Sublist.Common.csproj
@@ -109,6 +109,7 @@
   <ItemGroup>
     <Compile Include="Extensions\EnumerableExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Useful\PropertyChangedBase.cs" />
     <Content Include="Properties\Sublist.Common.rd.xml" />
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
