+++ Sublist.Implementation/Sublist.Implementation.csproj
@@ -108,13 +108,16 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App\AppData.cs" />
-    <Compile Include="Base\PropertyChangedBase.cs" />
     <Compile Include="Entries\SublistEntry.cs" />
     <Compile Include="Entries\SublistEntryBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Content Include="Properties\Sublist.Implementation.rd.xml" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\Sublist.Common\Sublist.Common.csproj">
+      <Project>{bce23953-f9f2-402a-8530-143101cb2cd8}</Project>
+      <Name>Sublist.Common</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Sublist.Contracts\Sublist.Contracts.csproj">
       <Project>{0c020dcb-611e-44c7-a5d0-619946528f02}</Project>
       <Name>Sublist.Contracts</Name>
